Sync w/ -current. 34 merge conflicts to follow.

+++ usr.bin/top/display.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: display.h,v 1.8 2007/05/24 20:04:04 ad Exp $	*/
+/*	$NetBSD: display.h,v 1.8.12.1 2008/06/23 04:32:11 wrstuden Exp $	*/
 
 /*
  *  Top users/processes display for Unix
@@ -62,5 +62,5 @@ void display_header __P((int));
 void new_message __P((int, const char *, ...))
      __attribute__((__format__(__printf__, 2, 3)));
 void clear_message __P((void));
-int readline __P((char *, int, int));
+double readline __P((char *, int, int));
 char *printable __P((char *));
